+++ structs.go
@@ -119,6 +119,30 @@ type Transfer struct {
 	Signature Signature `json:"signature" ssz-size:"96"`
 }
 
+type BeaconStatePhase0 struct {
+	GenesisTime                 uint64                `json:"genesis_time"`
+	GenesisValidatorsRoot       [32]byte              `json:"genesis_validators_root" ssz-size:"32"`
+	Slot                        uint64                `json:"slot"`
+	Fork                        *Fork                 `json:"fork"`
+	LatestBlockHeader           *BeaconBlockHeader    `json:"latest_block_header"`
+	BlockRoots                  [8192][32]byte        `json:"block_roots" ssz-size:"8192,32"`
+	StateRoots                  [8192][32]byte        `json:"state_roots" ssz-size:"8192,32"`
+	HistoricalRoots             [][32]byte            `json:"historical_roots" ssz-max:"16777216" ssz-size:"?,32"`
+	Eth1Data                    *Eth1Data             `json:"eth1_data"`
+	Eth1DataVotes               []*Eth1Data           `json:"eth1_data_votes" ssz-max:"2048"`
+	Eth1DepositIndex            uint64                `json:"eth1_deposit_index"`
+	Validators                  []*Validator          `json:"validators" ssz-max:"1099511627776"`
+	Balances                    []uint64              `json:"balances" ssz-max:"1099511627776"`
+	RandaoMixes                 [65536][32]byte       `json:"randao_mixes" ssz-size:"65536,32"`
+	Slashings                   []uint64              `json:"slashings" ssz-size:"8192"`
+	PreviousEpochAttestations   []*PendingAttestation `json:"previous_epoch_attestations" ssz-max:"4096"`
+	CurrentEpochAttestations    []*PendingAttestation `json:"current_epoch_attestations" ssz-max:"4096"`
+	JustificationBits           [1]byte               `json:"justification_bits" ssz-size:"1"`
+	PreviousJustifiedCheckpoint *Checkpoint           `json:"previous_justified_checkpoint"`
+	CurrentJustifiedCheckpoint  *Checkpoint           `json:"current_justified_checkpoint"`
+	FinalizedCheckpoint         *Checkpoint           `json:"finalized_checkpoint"`
+}
+
 type SignedBeaconBlockPhase0 struct {
 	Block     *BeaconBlockPhase0 `json:"message"`
 	Signature Signature          `json:"signature" ssz-size:"96"`
@@ -168,6 +192,33 @@ type SigningData struct {
 
 // Altair fork
 
+type BeaconStateAltair struct {
+	GenesisTime                 uint64             `json:"genesis_time"`
+	GenesisValidatorsRoot       [32]byte           `json:"genesis_validators_root" ssz-size:"32"`
+	Slot                        uint64             `json:"slot"`
+	Fork                        *Fork              `json:"fork"`
+	LatestBlockHeader           *BeaconBlockHeader `json:"latest_block_header"`
+	BlockRoots                  [8192][32]byte     `json:"block_roots" ssz-size:"8192,32"`
+	StateRoots                  [8192][32]byte     `json:"state_roots" ssz-size:"8192,32"`
+	HistoricalRoots             [][32]byte         `json:"historical_roots" ssz-max:"16777216" ssz-size:"?,32"`
+	Eth1Data                    *Eth1Data          `json:"eth1_data"`
+	Eth1DataVotes               []*Eth1Data        `json:"eth1_data_votes" ssz-max:"2048"`
+	Eth1DepositIndex            uint64             `json:"eth1_deposit_index"`
+	Validators                  []*Validator       `json:"validators" ssz-max:"1099511627776"`
+	Balances                    []uint64           `json:"balances" ssz-max:"1099511627776"`
+	RandaoMixes                 [65536][32]byte    `json:"randao_mixes" ssz-size:"65536,32"`
+	Slashings                   []uint64           `json:"slashings" ssz-size:"8192"`
+	PreviousEpochParticipation  []byte             `json:"previous_epoch_participation" ssz-max:"1099511627776"`
+	CurrentEpochParticipation   []byte             `json:"current_epoch_participation" ssz-max:"1099511627776"`
+	JustificationBits           [1]byte            `json:"justification_bits" ssz-size:"1"`
+	PreviousJustifiedCheckpoint *Checkpoint        `json:"previous_justified_checkpoint"`
+	CurrentJustifiedCheckpoint  *Checkpoint        `json:"current_justified_checkpoint"`
+	FinalizedCheckpoint         *Checkpoint        `json:"finalized_checkpoint"`
+	InactivityScores            []uint64           `json:"inactivity_scores" ssz-max:"1099511627776"`
+	CurrentSyncCommittee        *SyncCommittee     `json:"current_sync_committee"`
+	NextSyncCommittee           *SyncCommittee     `json:"next_sync_committee"`
+}
+
 type SignedBeaconBlockAltair struct {
 	Block     *BeaconBlockAltair `json:"message"`
 	Signature Signature          `json:"signature" ssz-size:"96"`
@@ -204,6 +255,34 @@ type SyncCommittee struct {
 }
 
 // bellatrix
+
+type BeaconStateBellatrix struct {
+	GenesisTime                  uint64                  `json:"genesis_time"`
+	GenesisValidatorsRoot        [32]byte                `json:"genesis_validators_root" ssz-size:"32"`
+	Slot                         uint64                  `json:"slot"`
+	Fork                         *Fork                   `json:"fork"`
+	LatestBlockHeader            *BeaconBlockHeader      `json:"latest_block_header"`
+	BlockRoots                   [8192][32]byte          `json:"block_roots" ssz-size:"8192,32"`
+	StateRoots                   [8192][32]byte          `json:"state_roots" ssz-size:"8192,32"`
+	HistoricalRoots              [][]byte                `json:"historical_roots" ssz-max:"16777216" ssz-size:"?,32"`
+	Eth1Data                     *Eth1Data               `json:"eth1_data"`
+	Eth1DataVotes                []*Eth1Data             `json:"eth1_data_votes" ssz-max:"2048"`
+	Eth1DepositIndex             uint64                  `json:"eth1_deposit_index"`
+	Validators                   []*Validator            `json:"validators" ssz-max:"1099511627776"`
+	Balances                     []uint64                `json:"balances" ssz-max:"1099511627776"`
+	RandaoMixes                  [65536][32]byte         `json:"randao_mixes" ssz-size:"65536,32"`
+	Slashings                    []uint64                `json:"slashings" ssz-size:"8192"`
+	PreviousEpochParticipation   []byte                  `json:"previous_epoch_participation" ssz-max:"1099511627776"`
+	CurrentEpochParticipation    []byte                  `json:"current_epoch_participation" ssz-max:"1099511627776"`
+	JustificationBits            [1]byte                 `json:"justification_bits" ssz-size:"1"`
+	PreviousJustifiedCheckpoint  *Checkpoint             `json:"previous_justified_checkpoint"`
+	CurrentJustifiedCheckpoint   *Checkpoint             `json:"current_justified_checkpoint"`
+	FinalizedCheckpoint          *Checkpoint             `json:"finalized_checkpoint"`
+	InactivityScores             []uint64                `json:"inactivity_scores" ssz-max:"1099511627776"`
+	CurrentSyncCommittee         *SyncCommittee          `json:"current_sync_committee"`
+	NextSyncCommittee            *SyncCommittee          `json:"next_sync_committee"`
+	LatestExecutionPayloadHeader *ExecutionPayloadHeader `json:"latest_execution_payload_header"`
+}
 
 type SignedBeaconBlockBellatrix struct {
 	Block     *BeaconBlockBellatrix `json:"message"`