+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.1.1 (Unreleased)
 
-- Add subscription endpoints for the beacon validator namespace [[GH-3](https://github.com/umbracle/eth2-validator/issues/3)]
+- feat: return wrap status code errors in Http client [[GH-8](https://github.com/umbracle/eth2-validator/issues/8)]
+- feat: `BeaconState` structs for `phase0`, `altair` and `bellatrix` forks [[GH-5](https://github.com/umbracle/eth2-validator/issues/5)]
+- feat: subscription endpoints for the beacon validator namespace [[GH-3](https://github.com/umbracle/eth2-validator/issues/3)]
 
 # 0.1.0 (29 July, 2022)